laboratoryGrade creation fix

+++ exams-app/src/app/examenes-improvisacion.service.ts
@@ -301,32 +301,37 @@ curl -m 70 -X POST https://us-central1-thoth-qa.cloudfunctions.net/deleteCertifi
   }  
 
   fileLoaded(db_collection, id, e:FileLoadedEvent){
-    db.collection(db_collection).doc(id).get().then( data =>{
-      //first erase the old value if existed
+    db.collection(db_collection).doc(id).get().then( doc =>{
       var promises = []
-      var oldFilePath:string =data[e.property + "Path"]
-      if( oldFilePath && oldFilePath != e.fileFullPath){
-        var storageOldRef = storage.ref( oldFilePath )
-        
-        var promiseDelete = storageOldRef.delete().then( () =>{
-          console.log("old file was deleted:" + oldFilePath )
-        })
-        .catch( reason => {
-          console.log("old file could not be deleted")      
-        })
-        promises.push(promiseDelete)
-      }  
-      //now update the values of properties to kick a reload of the data page
-      var values = {}
-      values[e.property + "Path"]=null                       
-      values[e.property + "Url"]=null
-      var remove = db.collection(db_collection).doc(id).update(values).then( () =>{
-        console.log("property has been update:" + e.property + " " + e.fileFullPath)
-      },
-      reason =>{
-        alert("ERROR: writing property:" + reason)
-      }) 
-      promises.push( remove )
+      var docExists = false
+      if( doc.exists ){
+        //first erase the old value if existed
+        docExists = true
+
+        var oldFilePath:string =doc[e.property + "Path"]
+        if( oldFilePath && oldFilePath != e.fileFullPath){
+          var storageOldRef = storage.ref( oldFilePath )
+          
+          var promiseDelete = storageOldRef.delete().then( () =>{
+            console.log("old file was deleted:" + oldFilePath )
+          })
+          .catch( reason => {
+            console.log("old file could not be deleted")      
+          })
+          promises.push(promiseDelete)
+        }  
+        //now update the values of properties to kick a reload of the data page
+        var values = {}
+        values[e.property + "Path"]=null                       
+        values[e.property + "Url"]=null
+        var remove = db.collection(db_collection).doc(id).update(values).then( () =>{
+          console.log("property has been update:" + e.property + " " + e.fileFullPath)
+        },
+        reason =>{
+          alert("ERROR: writing property:" + reason)
+        }) 
+        promises.push( remove )
+      }
 
       Promise.all( promises ).then( () =>{
         //now update the values in materia
@@ -334,13 +339,23 @@ curl -m 70 -X POST https://us-central1-thoth-qa.cloudfunctions.net/deleteCertifi
         storageRef.getDownloadURL().then( url =>{
             var values = {}
             values[e.property + "Path"]=e.fileFullPath                       
-            values[e.property + "Url"]=url        
-            db.collection(db_collection).doc(id).update(values).then( () =>{
-              console.log("property has been update:" + e.property + " " + e.fileFullPath)
-            },
-            reason =>{
-              alert("ERROR: writing property:" + reason)
-            })        
+            values[e.property + "Url"]=url     
+            if( docExists == true ){   
+              db.collection(db_collection).doc(id).update(values).then( () =>{
+                console.log("property has been update:" + e.property + " " + e.fileFullPath)
+              },
+              reason =>{
+                alert("ERROR: writing property:" + reason)
+              })        
+            }
+            else{
+              db.collection(db_collection).doc(id).set(values).then( () =>{
+                console.log("property has been set:" + e.property + " " + e.fileFullPath)
+              },
+              reason =>{
+                alert("ERROR: writing property:" + reason)
+              })                  
+            }  
         })       
       })
     })
